requirements.txt
jinja2>=3.1.0
weasyprint>=60.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
pyppeteer>=1.0.2 